Reduce public visibility (#312)
* Turns public modules to private or pub crate if possible
* Removes some unused code/macros
