+++ .github/scripts/ci-doc.sh
@@ -1,7 +1,11 @@
 . $(dirname "$0")/ci-common.sh
 
 # Check cargo doc
-cargo doc --features $non_exclusive_features --no-deps
+# We generate docs including private items so it would be easier for MMTk developers (GC implementers). However,
+# this could be confusing to MMTk users (binding implementers), as they may find items in the doc which
+# are not visible to a binding. If we exclude private items, the doc would be easier for the users, but would hide
+# implementation details for developers.
+cargo doc --features $non_exclusive_features --no-deps --document-private-items
 
 # Check tutorial code
 tutorial_code_dir=$project_root/docs/tutorial/code/mygc_semispace
