+++ src/mmtk.rs
@@ -1,6 +1,7 @@
+///! MMTk instance.
 use crate::plan::Plan;
 use crate::policy::space::SFTMap;
-use crate::scheduler::Scheduler;
+use crate::scheduler::MMTkScheduler;
 use crate::util::finalizable_processor::FinalizableProcessor;
 use crate::util::heap::layout::heap_layout::Mmapper;
 use crate::util::heap::layout::heap_layout::VMMap;
@@ -24,38 +25,37 @@ lazy_static! {
     // 2. These mmappers are possibly global across multiple MMTk instances, as they manage the
     //    entire address space.
     // TODO: We should refactor this when we know more about how multiple MMTK instances work.
+
+    /// A global VMMap that manages the mapping of spaces to virtual memory ranges.
     pub static ref VM_MAP: VMMap = VMMap::new();
+
+    /// A global Mmapper for mmaping and protection of virtual memory.
     pub static ref MMAPPER: Mmapper = Mmapper::new();
+
+    // A global space function table that allows efficient dispatch space specific code for addresses in our heap.
     pub static ref SFT_MAP: SFTMap<'static> = SFTMap::new();
 }
 
-/// An MMTk instance. MMTk allows mutiple instances to run independently, and each instance gives users a separate heap.
+/// An MMTk instance. MMTk allows multiple instances to run independently, and each instance gives users a separate heap.
 /// *Note that multi-instances is not fully supported yet*
 pub struct MMTK<VM: VMBinding> {
-    pub plan: Box<dyn Plan<VM = VM>>,
-    pub vm_map: &'static VMMap,
-    pub mmapper: &'static Mmapper,
-    pub sftmap: &'static SFTMap<'static>,
-    pub reference_processors: ReferenceProcessors,
-    pub finalizable_processor: Mutex<FinalizableProcessor>,
-    pub options: Arc<UnsafeOptionsWrapper>,
-    pub scheduler: Arc<Scheduler<Self>>,
+    pub(crate) plan: Box<dyn Plan<VM = VM>>,
+    pub(crate) reference_processors: ReferenceProcessors,
+    pub(crate) finalizable_processor: Mutex<FinalizableProcessor>,
+    pub(crate) options: Arc<UnsafeOptionsWrapper>,
+    pub(crate) scheduler: Arc<MMTkScheduler<VM>>,
     #[cfg(feature = "sanity")]
-    pub sanity_checker: Mutex<SanityChecker>,
+    pub(crate) sanity_checker: Mutex<SanityChecker>,
     inside_harness: AtomicBool,
 }
 
 impl<VM: VMBinding> MMTK<VM> {
     pub fn new() -> Self {
-        let scheduler = Scheduler::new();
+        let scheduler = MMTkScheduler::new();
         let options = Arc::new(UnsafeOptionsWrapper::new(Options::default()));
-        let plan =
-            crate::plan::global::create_plan(options.plan, &VM_MAP, &MMAPPER, options.clone());
+        let plan = crate::plan::create_plan(options.plan, &VM_MAP, &MMAPPER, options.clone());
         MMTK {
             plan,
-            vm_map: &VM_MAP,
-            mmapper: &MMAPPER,
-            sftmap: &SFT_MAP,
             reference_processors: ReferenceProcessors::new(),
             finalizable_processor: Mutex::new(FinalizableProcessor::new()),
             options,
@@ -77,6 +77,10 @@ impl<VM: VMBinding> MMTK<VM> {
     pub fn harness_end(&'static self) {
         self.plan.base().stats.stop_all(self);
         self.inside_harness.store(false, Ordering::SeqCst);
+    }
+
+    pub fn get_plan(&self) -> &dyn Plan<VM = VM> {
+        self.plan.as_ref()
     }
 }
