+++ src/plan/marksweep/mod.rs
@@ -1,3 +1,5 @@
+//! Plan: marksweep (currently using malloc as its freelist allocator)
+
 mod gc_work;
 mod global;
 pub mod mutator;
