+++ src/policy/copyspace.rs
@@ -16,6 +16,7 @@ use std::sync::atomic::{AtomicBool, Ordering};
 
 const META_DATA_PAGES_PER_REGION: usize = CARD_META_PAGES_PER_REGION;
 
+/// This type implements a simple copying space.
 pub struct CopySpace<VM: VMBinding> {
     common: CommonSpace<VM>,
     pr: MonotonePageResource<VM>,
@@ -153,6 +154,7 @@ impl<VM: VMBinding> CopySpace<VM> {
         }
     }
 
+    #[allow(dead_code)] // Only used with certain features (such as sanity)
     pub fn protect(&self) {
         if !self.common().contiguous {
             panic!(
@@ -167,6 +169,7 @@ impl<VM: VMBinding> CopySpace<VM> {
         trace!("Protect {:x} {:x}", start, start + extent);
     }
 
+    #[allow(dead_code)] // Only used with certain features (such as sanity)
     pub fn unprotect(&self) {
         if !self.common().contiguous {
             panic!(
