+++ src/policy/mallocspace/mod.rs
@@ -1,3 +1,4 @@
+///! A marksweep space that allocates from malloc.
 mod global;
 pub mod metadata;
