+++ vmbindings/dummyvm/src/scanning.rs
@@ -2,8 +2,7 @@ use mmtk::vm::Scanning;
 use mmtk::{TransitiveClosure, Mutator};
 use mmtk::util::ObjectReference;
 use mmtk::util::opaque_pointer::*;
-use mmtk::scheduler::gc_work::*;
-use mmtk::scheduler::GCWorker;
+use mmtk::scheduler::*;
 use crate::DummyVM;
 
 pub struct VMScanning {}
